requirements.txt
python-dotenv~=1.0.1
mysql-connector-python~=9.0.0
requests~=2.31.0
beautifulsoup4~=4.12.3
selenium~=4.19.0
webdriver-manager~=4.0.1
spotipy~=2.23.0
google-api-python-client~=2.125.0